bug: DMS V2 Float should be float64 [DX-1234] (#881)

+++ libs/platypus-core/src/lib/domain/data-model/constants.ts
@@ -6,7 +6,7 @@ export const mixerApiBuiltInTypes = [
   { name: 'String', type: 'SCALAR', dmsType: 'text' },
   { name: 'Int', type: 'SCALAR', dmsType: 'int32' },
   { name: 'Int64', type: 'SCALAR', dmsType: 'int64' },
-  { name: 'Float', type: 'SCALAR', dmsType: 'float32' },
+  { name: 'Float', type: 'SCALAR', dmsType: 'float64' },
   { name: 'Boolean', type: 'SCALAR', dmsType: 'boolean' },
   { name: 'Timestamp', type: 'SCALAR', dmsType: 'timestamp' },
   { name: 'JSONObject', type: 'SCALAR', dmsType: 'json' },
